okokok

+++ funcao-perimetro.cpp
@@ -1,14 +1,14 @@
 #include <iostream>
 
-int perimetro(int a, int b){
-    int r = a *2+ b*2;
+long long perimetro(int a, int b){
+    long long r = a * 2LL + b* 2LL;
     return r;
 }
 
 int main(){
     int a,b;
     std::cin >> a >> b;
-    int p = perimetro(a,b);
+    long long p = perimetro(a,b);
     std::cout << p << std::endl;
     return 0;
 }
